% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Com_Media.R
\name{Come}
\alias{Come}
\title{Comparación de Medias}
\usage{
Come(A, B, C)
}
\arguments{
\item{A}{(Vector) tratamiento 1.}

\item{B}{(Vector) tratamiento 2.}

\item{C}{(Vcetor) tratamiento 3.}

\item{df}{base de datos}
}
\description{
Realiza una tabla ANOVA para la comparación de medias de tres tratamientos.
}
\examples{
/dontrun{
##Directorio de trabajo
ruta <- "~/ruta/test.csv"

------------------------------
#Ejemplo 1
df <- read.csv(ruta)
# Cargo la libreria
library(ComMedia)
A, B Y C representan los tres tratamientos
Com_Media(A = df$A, B = df$B, C = df$C)

------------------------------
# Ejemplo 2
A <- c(39.5,41.2,38.8,42.1,40.4,37.9,44.2,42.7,36.5,45.8,40.1,43.6,39.9,42.3,38.2,46.5,41.7,38.9,47.9,40.8)
B <- c(42.1,44.3,41.8,45.6,43.9,40.5,47.2,46.1,39.7,48.8,43.2,46.9,42.6,45.3,41.4,48.2,44.6,41.9,49.7,43.8)
c <- c(45.2,47.5,46.8,44.3,46.6,43.9,51.1,48.7,42.4,50.845.9,48.2,46.6,49.1,44.8,52.6,49.3,47.1,53.4,46.7)

# Cargo la función
Com_Media(A = A, B = B, C = C)
}
}
